NGSTACK-489 Implement content and location field mapper
