+++ tests/lib/Integration/Implementation/Solr/FieldMapper/TestFieldMapperInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Netgen\EzPlatformSearchExtra\Tests\Integration\Implementation\Solr\FieldMapper;
+
+interface TestFieldMapperInterface
+{
+    const CONTENT_TYPE_IDENTIFIER = 'extra_fields_test';
+
+    const CHILD_CONTENT_TYPE_IDENTIFIER = 'extra_fields_test_comment';
+}
